services/api-gateway/src/services/python/requirements.txt
inspireface
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.0.0
